Change type

+++ ShuffleSource.cpp
@@ -2,7 +2,7 @@
 #include <math.h>
 #include <random>
 
-uint8_t map(size_t num, size_t bitDepth) {
+uint8_t map(uint32_t num, uint32_t bitDepth) {
     num = (num + 1) << (8 - bitDepth);
     return static_cast<uint8_t>(num - 1);
 }
@@ -11,7 +11,7 @@ ShuffleSource::ShuffleSource(const Options& options) {
     std::vector<Color32> colors;
     uint32_t bitDepth = options.bitDepth;
 
-    size_t max = static_cast<size_t>(pow(2, bitDepth));
+    uint32_t max = static_cast<size_t>(pow(2, bitDepth));
     for (size_t r = 0; r < max; r++) {
         for (size_t g = 0; g < max; g++) {
             for (size_t b = 0; b < max; b++) {
